feat: moved conceptually core files to a 'core' folder; updated imports
