+++ .gitignore
@@ -1,3 +1,4 @@
+.python-version
 db/
 ./source_docs/*
 tests/mocks/db/*
