+++ lib/core/pyproject.toml
@@ -22,6 +22,7 @@ pydantic = "^2.10.4"
 requires = ["poetry-core"]
 build-backend = "poetry.core.masonry.api"
 
+
 [tool.mypy]
 python_version = "3.13"
 strict = true
