+++ lib/core/view_model.py
