+++ tests/types/test_abstractclasses_subtypes.py
@@ -9,9 +9,9 @@ Have in mind that not all tests use the --strict flag
 from abc import abstractmethod
 import sys
 from typing import Generic, NewType, TypeVar
-from lib.bac import BaseAbstractClass
-from lib.primary_ports import BaseInputPort, BaseOutputPort
-from lib.usecase_models import BaseErrorResponseModel, BaseRequestModel, BaseResponseModel
+from lib.core.bac import BaseAbstractClass
+from lib.core.primary_ports import BaseInputPort, BaseOutputPort
+from lib.core.usecase_models import BaseErrorResponseModel, BaseRequestModel, BaseResponseModel
 
 from tests.types.mypytest_prototype import debug_mypy, object_mypy_error_report
